Reorganize components into AboutPage folder
