+++ src/v2/components/FeatureSlides/slides.js
