+++ src/v2/pages/about/AboutPage/index.tsx
@@ -2,13 +2,13 @@ import React from 'react'
 import styled from 'styled-components'
 
 import Box from 'v2/components/UI/Box'
-import { TopLogin } from '../components/TopLogin'
+import { TopLogin } from 'v2/components/TopLogin'
 import BlankLayout from 'v2/components/UI/Layouts/BlankLayout'
 import Text from 'v2/components/UI/Text'
-import { P } from '../components/Common'
-import { calculateAge } from '../lib/calculateAge'
-import FeatureCarouselWithSlides from '../components/FeatureSlides'
-import PricingTable from '../components/PricingTable'
+import { P } from 'v2/pages/about/components/Text'
+import { calculateAge } from 'v2/util/calculateAge'
+import FeatureCarouselWithSlides from 'v2/components/FeatureSlides'
+import PricingTable from 'v2/pages/about/PricingPage/components/PricingTable'
 import { RoadmapPageInner } from 'v2/pages/about/RoadmapPage'
 import { EssaysCommunity } from './components/EssaysCommunity'
 import { PricingQuestions } from './components/PricingQuestions'
