+++ src/v2/pages/about/components/Text/index.tsx
@@ -34,4 +34,18 @@ const Description = styled(Text).attrs({
   }
 `
 
+export const P = styled(Text).attrs({
+  f: 4,
+  color: 'gray.bold',
+  lineHeight: 1.5,
+  mb: 7,
+})`
+  text-align: left;
+
+  ul {
+    margin-left: 0;
+    padding-left: ${x => x.theme.space[7]};
+  }
+`
+
 export { Headline, Subheadline, Description }
