+++ src/v2/pages/home/HomePage/index.tsx
@@ -4,9 +4,9 @@ import styled from 'styled-components'
 import Box from 'v2/components/UI/Box'
 import BlankLayout from 'v2/components/UI/Layouts/BlankLayout'
 // import { DesireLine } from './components/DesireLine'
-import FeatureCarouselWithSlides from 'v2/pages/home/components/FeatureSlides'
+import FeatureCarouselWithSlides from 'v2/components/FeatureSlides'
 import { IntroParagraph } from './components/IntroParagraph'
-import { TopLogin } from 'v2/pages/home/components/TopLogin'
+import { TopLogin } from 'v2/components/TopLogin'
 import { LoggedOutFooter } from 'v2/components/LoggedOutFooter'
 import { TheSecret } from './components/TheSecret'
 import { LongTermVision } from 'v2/pages/home/components/LongTermVision'
